tool/requirements.txt
Pillow
strictyaml